EOS-19662: Fixed review comments
